move more methods to harmony
TollBooth.EnterBuildingSegment to harmony patch
EconomyManager.AddPrivateIncome use redirection
